+++ init.rb
@@ -29,7 +29,7 @@ Redmine::Plugin.register :redmine_xapian do
   author_url 'https://github.com/xelkano/redmine_xapian/graphs/contributors'
 
   description 'With this plugin you will be able to do searches by file name and by strings inside your documents'
-  version '3.0.3'
+  version '3.0.4 devel'
   requires_redmine version_or_higher: '4.1.0'
 
   languages = %w[danish dutch english finnish french german german2 hungarian italian kraaij_pohlmann